requirement.txt
langchain
streamlit
python-dotenv
langchain-groq
google-generativeai
faiss-cpu
pypdf
langchain_huggingface